requirements.txt
numpy==1.26.3
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==1.4.46
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0